backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.27
asyncpg>=0.29.0
alembic>=1.13.0
google-generativeai>=0.4.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
supabase>=2.3.0
python-multipart>=0.0.9
structlog>=24.1.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
factory-boy>=3.3.0
sentry-sdk[fastapi]>=1.41.0
